+++ client/jackson/build.gradle
@@ -1,0 +1,14 @@
+apply plugin: 'java'
+apply plugin: 'kotlin'
+apply plugin: 'net.corda.plugins.publish-utils'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    compile project(':core')
+    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    compile ("com.fasterxml.jackson.module:jackson-module-kotlin:${jackson_version}")
+    testCompile "junit:junit:$junit_version"
+}
