+++ samples/irs-demo/src/main/kotlin/net/corda/simulation/IRSSimulation.kt
@@ -18,13 +18,11 @@ import net.corda.flows.TwoPartyDealFlow.Acceptor
 import net.corda.flows.TwoPartyDealFlow.AutoOffer
 import net.corda.flows.TwoPartyDealFlow.Instigator
 import net.corda.irs.contract.InterestRateSwap
+import net.corda.jackson.JacksonSupport
 import net.corda.node.utilities.databaseTransaction
 import net.corda.testing.initiateSingleShotFlow
 import net.corda.testing.node.InMemoryMessagingNetwork
 import net.corda.testing.node.MockIdentityService
-import net.i2p.crypto.eddsa.KeyPairGenerator
-import java.security.SecureRandom
-import net.corda.webserver.utilities.JsonSupport
 import java.time.LocalDate
 import java.util.*
 
@@ -33,7 +31,7 @@ import java.util.*
  * A simulation in which banks execute interest rate swaps with each other, including the fixing events.
  */
 class IRSSimulation(networkSendManuallyPumped: Boolean, runAsync: Boolean, latencyInjector: InMemoryMessagingNetwork.LatencyCalculator?) : Simulation(networkSendManuallyPumped, runAsync, latencyInjector) {
-    val om = JsonSupport.createInMemoryMapper(MockIdentityService(network.identities))
+    val om = JacksonSupport.createInMemoryMapper(MockIdentityService(network.identities))
 
     init {
         currentDateAndTime = LocalDate.of(2016, 3, 8).atStartOfDay()
