+++ settings.gradle
@@ -9,6 +9,7 @@ include 'node'
 include 'node:capsule'
 include 'node:webserver'
 include 'client'
+include 'client:jackson'
 include 'experimental'
 include 'experimental:sandbox'
 include 'test-utils'
@@ -22,4 +23,3 @@ include 'samples:network-visualiser'
 include 'samples:simm-valuation-demo'
 include 'samples:raft-notary-demo'
 include 'samples:bank-of-corda-demo'
-
